+++ packages/esp_board_manager/test_apps/main/periph/test_periph_i2c.c
@@ -1,5 +1,6 @@
 #include <stdio.h>
 #include "esp_board_periph.h"
+#include "esp_board_manager_defs.h"
 #include "driver/i2c_master.h"
 #include "sdmmc_cmd.h"
 #include "driver/sdmmc_host.h"
@@ -8,7 +9,7 @@
 void test_periph_i2c(void)
 {
     /* Initialize I2C peripheral */
-    esp_err_t ret = esp_board_periph_init("i2c_master");
+    esp_err_t ret = esp_board_periph_init(ESP_BOARD_PERIPH_NAME_I2C_MASTER);
     if (ret != ESP_OK) {
         printf("Failed to initialize I2C master peripheral\n");
         return;
@@ -16,7 +17,7 @@ void test_periph_i2c(void)
 
     /* Get I2C handle */
     void *i2c_handle = NULL;
-    ret = esp_board_periph_get_handle("i2c_master", &i2c_handle);
+    ret = esp_board_periph_get_handle(ESP_BOARD_PERIPH_NAME_I2C_MASTER, &i2c_handle);
     if (ret != ESP_OK || !i2c_handle) {
         printf("Failed to get I2C master handle\n");
         goto cleanup;
@@ -49,10 +50,10 @@ void test_periph_i2c(void)
     i2c_master_bus_rm_device(dev_handle);
 
     /* Show peripheral information */
-    esp_board_periph_show("i2c_master");
+    esp_board_periph_show(ESP_BOARD_PERIPH_NAME_I2C_MASTER);
 
 cleanup:
     /* Cleanup */
-    esp_board_periph_deinit("i2c_master");
+    esp_board_periph_deinit(ESP_BOARD_PERIPH_NAME_I2C_MASTER);
     printf("I2C master peripheral deinitialized\n");
 }